Make the broker manage its own sidecars.
 * Jezebel "working" resides in extra-{modules,managed} completely.
 * Backtrace (coronerd) will now be managed by noitd if MANAGED_CORONER
   is set in the environment.

Ideally we'll ship jezebel, jars and its dependency on a JRE as a
separate package (which is installed by default on the broker).  With
this setup, we can now more easily manage a docker image as `noitd` is
a sufficient process to start as it will make sure that crash reporting
and jezebel are operating.