Fix docker upgrades and volume interplay.
If /opt/noit/prod/etc is a volume, we get the "original" installs
copy in there and it will never upgrade.  To work around that we
stash a copy of the deb used to install these files in /opt/napp/package
and if it doesn't match the version last used in the volume we force
a reinstall (which don't touch the user configs) and mark it as done.
This should happen once per upgrade.