Update TLS settings on default listener
Permit only TLS 1.2 and later, and reduce the cipher set to GCM
suites, which have higher performance. The vast majority of clients
are already using GCM ciphers, and with the removal of older protocol
versions, there is no need to support less secure/performant ciphers.

Update example in listener site config to reflect re-enabling older
TLS versions for an alternate port, as that is the most likely scenario.